+++ CHANGELOG.md
@@ -14,6 +14,7 @@ ENHANCEMENTS:
 * resource/hopsworksai_cluster: Set Computed to true for `azure_attributes/network/subnet_name`,  `azure_attributes/network/security_group_name`, `aws_attributes/network/subnet_id` and `aws_attributes/network/security_group_id`
 * datasource/azure_user_assigned_identity_permissions: Add a new attribute `enable_aks_and_acr`
 * resource/hopsworksai_cluster: Set Computed to true for `aws_attributes/ecr_registry_account_id` 
+* complete example to create Hopsworks clusters with AKS/ACR and EKS/ECR
 
 FEATURES:
 * **New Data Source**: `hopsworksai_instance_type`